backend/python/requirements.txt
fastapi
uvicorn
torch
diffusers
transformers
accelerate
safetensors
opencv-python
pillow
controlnet-aux
xformers
requests
python-multipart
